tests: gather all basic ctypes tests in test_ctypes.py
Move various basic tests for `ctypes` into `test_ctypes.py`.
